requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
